backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Task queue
celery==5.4.0
redis==5.2.1

# HTTP clients
httpx==0.28.1
aiohttp==3.11.11

# LLM providers
openai==1.58.1
anthropic==0.40.0
google-generativeai==0.8.3

# Amazon SP-API
python-amazon-sp-api==1.0.6

# Web scraping
playwright==1.49.0
crawl4ai==0.4.21
beautifulsoup4==4.12.3
lxml==5.3.0

# Data processing
pandas==2.2.3
numpy==2.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Utilities
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
